requirements.txt
Appium-Python-Client==0.50
pytest==4.6.9
six==1.14.0